+++ src/funproxy/run.py
@@ -1,0 +1,4 @@
+from funproxy.job import GetFreeProxy
+
+if __name__ == "__main__":
+    GetFreeProxy().run(5)